# Optimization batch 8: use file basenames even more