… Documentation/gitdiffcore.txt
… diffcore.h
… merge-ort.c