This series depends on ????.

This series uses file basenames in a more involved fashion to guide
rename detection.  It's a follow-on improvement to "Optimization #3"
from my Git Merge 2020 talk[1].  For the testcases mentioned in commit
?????????? ("merge-ort: begin performance work; instrument with
trace2_region_* calls", 2020-10-28), the changes in just this series
improves the performance as follows:

                         Before Series           After Series
    no-renames:      13.138 s ±  0.086 s     12.544 s ±  0.069 s
    mega-renames:   169.488 s ±  0.494 s    123.506 s ±  0.372 s
    just-one-mega:    5.061 s ±  0.017 s      3.757 s ±  0.011 s

As a reminder, before any merge-ort/diffcore-rename performance work,
the performance results we started with (as noted in the same commit
message) were:

    no-renames-am:      6.940 s ±  0.485 s
    no-renames:        18.912 s ±  0.174 s
    mega-renames:    5964.031 s ± 10.459 s
    just-one-mega:    149.583 s ±  0.751 s


There was an alternative idea to the series presented here that I also
tried: instead of using directory rename detection based on exact
renames to predict where files would be renamed and then comparing to
the file with the same basename in the new directory, one could
instead take all files with the same basename -- both sources and
destinations -- and then do a smaller M x N comparison on all those
files to find renames.  Any non-matches after that step could be
combined with all other files for the big inexact rename detection
step.

There are two problems with such a strategy, though.

One is that in the worst case, you approximately double the cost of
rename detection (if most potential rename pairs all have the same
basename but they aren't actually matches, you end up comparing
twice).

The second issue isn't clear until trying to combine this idea with
later performance optimizations.  The next optimization will provide a
way to filter out several of the rename sources.  If our inexact
rename detection matrix is sized 1 x 4000 because we can remove all
but one source file, but we have 100 files with the same basename,
then a 100 x 100 comparison is actually more costly than a 1 x 4000
comparison -- and we don't need most of the renames from the 100 x 100
comparison.  The advantage of the directory rename detection based
idea for finding which basenames to match up, is that the cost for
each file is linear (or, said another way, scales proportional to
doing a diff on that file).  As such, the costs for this preliminary
optimization are nicely controlled and the worst case scenario is it
has spent a little extra time upfront but still has to do the full
inexact rename detection.

[1] https://github.com/newren/presentations/blob/pdfs/merge-performance/merge-performance-slides.pdf